feat: enable local dev publishing of documentation using docker
When developing documentation in standards-* repositories, there was no
easy way to view your documentation as you were working on it to ensure
it looked correct etc. This commit adds a container image definition
which enables users to do this in their own standards-* repositories.

Because users will use the image directly and mount their documentation
as a volume (doc in readme to come) into the running container, the
methods this repository uses when building the eleventy site (i.e. the
`obtain-docs.js` and `generate-docs-file-meta.js` scripts) aren't run
and therefore some minor tweaks had to be made to ensure the lack of a
git repository at the root of the /site directory in the container
didn't break everything. This is achieved by using an env var set in
the Dockerfile SKIP_META which if present, results in the meta about a
page not being generated using git and the meta.json files, but using
some sensible defaults.