build: add more linting for the dockerfile using hadolint
This adds an action workflow as well as information in the readme
about how to run hadolint locally using docker.

I originally looked for a javascript based dockerfile linter but
they all seemed very out of date so hadolint was the best bet.
There is a hadolint wrapper but the code wasn't available and it
has been sporadically updated and not very well used so I chose
not to use it to ensure we weren't introducing a bad dependency
into the repo where it's not super needed.