test: stop using buildkit for test image
We had a problem where seemingly the pull of the base debian image
used for the initial build step was failing (we'd get an error
when pulling the metadata: "context deadline exceeded"). Exactly
why this was a problem was unclear but it only happened when the
debian image wasn't available, if it was it worked. Removing
buildkit from the equation seems to solve this problem and I suspect
this may be related to why dockerode (the underlying docker lib
testcontainers uses on nodejs) doesn't use it by default. There are
some minor changes required in the dockerfile to make this work too.