feat: add a placeholder page for Ways of Working until the repo is ready
There's no repo yet and the doc isn't complete but we still have a link
to WoW documentation from the home page. Instead of just removing it from
the homepage, it feels better to add a placeholder page until it's ready.

We'll have to make sure that when we add the WoW configuration properly,
we remove the ./docs/wow dir and remove the temp .gitignore entry.